Sprint 16: the Sprint that Never Ended (#280)
* rebuild for updated IIIF API

ingest from DLXS + even more SEO friendly URLs

* update local master

* make selection modal

* tune selection and add home

* add updated home and search form

* fix search tracking

* track keycodes before canvas has focus

* fix landscape results page highlighting

* remove opensearch setup

* support more keyboard events

* labels for selects

* alt text update

* don't crash when solr is returning no highlighted text

* a11y outline

* embiggen experiment

* embiggen experiment

* embiggen mode creates the largest possible viewer based on window width

* keep embiggened between page turns

* fix pageviewer for embiggened

* issue titles

* add on focus/blur action to side graph

* fixup

* add obvious search link color and underline

* add sr-only labels to browse controls

* add accessible link color and underline for home and browse

* add correction to subittle color & underline

* fix views

* add delays to play nice with DLXS

* embiggen experiment

* embiggen experiment

* embiggen mode creates the largest possible viewer based on window width

* keep embiggened between page turns

* fix pageviewer for embiggened

* issue titles

* event hacking

* so many options

* sticky controls

* make option=tabs default

* minor cleanup

* make issue link a filter vs. query; fix the search form

* fix for graph bug goto to home page

* add delay

* add delays, progress, and support for cobbering

* better reporting

* graph but fix for previous bug fix

* option=tabbish

* tabbish responsive

* Image > Page Image

* Image > Page Image

* tweak mobile viewer

* remove redundant query parameters

* lazily load images

* tweak styles

* link to ocr help

* responsive tweaks to viewer

* maybe possibly fix highlight bug

* favor building current_search_session on params iff start_new_session? true

* support reading text files from filesystem

* remove delays

* yank more sleep

* make async an explicit use case for thumbnail src

* make option=stacked_tabs concrete; remove obsolete views

+++ app/services/dlxs_ingest.rb
@@ -44,15 +44,21 @@ module IIIF
   end
 
   class Service
-    attr_accessor :data
+    attr_accessor :data, :response
     def self.parse(response)
+      @response = response
       ob = Nop.new(JSON.parse(response))
+    rescue Exception => e
+      STDERR.puts "JSON PARSE ERROR: #{e}"
+      STDERR.puts "[#{@response}]"
+      raise
     end
   end
 end
 
 class DlxsIngest
 
+  attr_accessor :clobber
 
   def initialize(publication_slug, collid)
     @publication = Publication.find_by_slug(publication_slug)
@@ -67,9 +73,13 @@ class DlxsIngest
     tdelta = t0
     total = collection.collections.size
     collection.collections.each_with_index do |member, member_idx|
-      fetch_volume(member['@id'])
+      check = fetch_volume(member['@id'])
       STDERR.puts "-- #{Time.now - tdelta} : #{member_idx} / #{total} : #{member['@id']}"
       tdelta = Time.now
+      # unless check == false
+      #   delay = rand() * 10
+      #   sleep(delay)
+      # end
     end
     STDERR.puts "-- #{Time.now - t0} : EOT"
   end
@@ -78,25 +88,45 @@ class DlxsIngest
     volume_url = service_collection_url(volume_identifier)
     volume_collection = IIIF::Service.parse(fetch_data(volume_url))
 
+    unless @clobber
+      check = Issue.where(volume_identifier: local_identifier(volume_identifier)).first
+      STDERR.puts "== SKIPPING #{volume_identifier}" unless check.nil?
+      return false unless check.nil?
+    end
 
     # delete all the issues because the issue manifests may have changed
     Issue.where(volume_identifier: local_identifier(volume_identifier)).destroy_all
 
-    volume_collection.manifests.each do |member|
-      fetch_issue(member['@id'])
+    total = volume_collection.manifests.length
+    volume_collection.manifests.each_with_index do |member, idx|
+      fetch_issue(member['@id'], idx, total)
+      # delay = rand() * 5
+      # sleep(delay)
     end
   end
 
-  def fetch_issue(issue_identifier)
+  def fetch_issue(issue_identifier, idx=0, total=0)
     issue_url = service_manifest_url(issue_identifier)
     issue_manifest = IIIF::Service.parse(fetch_data(issue_url))
 
-    build_issue(issue_manifest)
+    build_issue(issue_manifest, idx, total)
+  rescue
+    STDERR.puts "FAILED #{issue_url}"
+    raise
   end
 
   def fetch_data(url)
     uri = URI(url)
-    response = Net::HTTP.get(uri)
+    response = Net::HTTP.get_response(uri)
+    unless response.is_a?(Net::HTTPSuccess)
+      STDERR.puts "FAILED: #{response.code}"
+      PP.pp response, STDERR
+      if response.is_a?(Net::HTTPRedirection) or response.is_a?(Net::HTTPFound)
+        STDERR.puts "JUMPING TO #{response['location']}"
+      end
+      return ''
+    end
+    response.body
   end
 
   def service_collection_url(identifier)
@@ -114,19 +144,19 @@ class DlxsIngest
     identifier # + "?prep=1"
   end
 
-  ## 
+  ##
 
-  def build_issue(manifest)
+  def build_issue(manifest, idx=0, total=0)
     metadata = build_metadata(manifest)
     issue = Issue.create \
       volume_identifier: metadata[:volume_identifier],
       issue_identifier: metadata[:issue_identifier],
-      volume: metadata[:issue_volume], 
-      issue_number: metadata[:issue_number], 
-      edition: metadata[:issue_edition], 
+      volume: metadata[:issue_volume],
+      issue_number: metadata[:issue_number],
+      edition: metadata[:issue_edition],
       date_issued: metadata[:date_issued],
       publication_year: Time.new(metadata[:date_issued]).strftime("%Y"),
-      variant_sequence: metadata[:variant_sequence].to_i, 
+      variant_sequence: metadata[:variant_sequence].to_i,
       publication_title: metadata[:publication_title],
       publication: @publication
 
@@ -140,7 +170,7 @@ class DlxsIngest
       end
     end
 
-    STDERR.puts "== #{issue.issue_identifier} : #{num_processed} : #{Time.now - t0}"
+    STDERR.puts "== #{idx} / #{total} : #{issue.issue_identifier} : #{num_processed} : #{Time.now - t0}"
 
   end
 
@@ -159,9 +189,9 @@ class DlxsIngest
       coordinates_link = File.basename(coordinates_link['@id'])
     end
 
+
     # image_link = File.basename(canvas.images.first.resource['@id'])
     image_link = File.basename(canvas.images.first['resource']['@id'])
-
     page = Page.create \
       issue: issue,
       height: canvas.height,
